fix(logger): launcher debug logs are not displayed by default when verbose is 4
