+++ packages/nlu/src/nlu-server/index.ts
@@ -2,44 +2,13 @@ import bytes from 'bytes'
 import chalk from 'chalk'
 import _ from 'lodash'
 import path from 'path'
-import * as NLUEngine from '../engine'
 import Logger, { centerText } from '../utils/logger'
-import { LoggerLevel, ILogger } from '../utils/logger/typings'
+import { LoggerLevel } from '../utils/logger/typings'
 import API from './api'
-import { CommandLineOptions, getConfig, StanOptions } from './config'
+import { CommandLineOptions, getConfig } from './config'
 import { copyDir } from './copy-dir'
 import { displayDocumentation } from './documentation'
-
-const makeEngine = async (options: StanOptions, logger: ILogger) => {
-  const loggerWrapper: NLUEngine.Logger = {
-    debug: (msg: string) => logger.debug(msg),
-    info: (msg: string) => logger.info(msg),
-    warning: (msg: string, err?: Error) => (err ? logger.attachError(err).warn(msg) : logger.warn(msg)),
-    error: (msg: string, err?: Error) => (err ? logger.attachError(err).error(msg) : logger.error(msg))
-  }
-
-  try {
-    const { ducklingEnabled, ducklingURL, modelCacheSize, languageSources, legacyElection } = options
-    const config: NLUEngine.Config = {
-      languageSources,
-      ducklingEnabled,
-      ducklingURL,
-      modelCacheSize,
-      legacyElection
-    }
-
-    const engine = await NLUEngine.makeEngine(config, loggerWrapper)
-    return engine
-  } catch (err) {
-    // TODO: Make lang provider throw if it can't connect.
-    logger
-      .attachError(err)
-      .error(
-        'There was an error while initializing Engine tools. Check out the connection to your language and Duckling server.'
-      )
-    process.exit(1)
-  }
-}
+import { makeEngine } from './make-engine'
 
 export default async function (cliOptions: CommandLineOptions, version: string) {
   const { options, source: configSource } = await getConfig(cliOptions)
@@ -50,10 +19,8 @@ export default async function (cliOptions: CommandLineOptions, version: string) 
   })
 
   const launcherLogger = Logger.sub('Launcher')
-  // Launcher always display
   launcherLogger.configure({
-    level: Math.max(options.verbose, LoggerLevel.Info),
-    filters: ['']
+    minLevel: LoggerLevel.Info // Launcher always display
   })
 
   for (const dir of ['./pre-trained', './stop-words']) {
